Privatise + rename CompositeDrawableDrawNode
